+++ packages/area/index.ts
@@ -1,15 +1,19 @@
 import { VantComponent } from '../common/component';
 
 type AreaItem = {
-  name: string;
-  code: string;
+  name: string
+  code: string
 };
 
 VantComponent({
+  classes: ['active-class', 'toolbar-class', 'column-class'],
+
   props: {
     title: String,
     value: String,
     loading: Boolean,
+    cancelButtonText: String,
+    confirmButtonText: String,
     itemHeight: {
       type: Number,
       value: 44
@@ -29,62 +33,62 @@ VantComponent({
   },
 
   data: {
-    pickerValue: [0, 0, 0],
-    columns: []
-  },
-
-  computed: {
-    displayColumns() {
-      const { columns = [], columnsNum } = this.data;
-      return columns.slice(0, +columnsNum);
-    }
+    columns: [{ values: [] }, { values: [] }, { values: [] }],
+    displayColumns: [{ values: [] }, { values: [] }, { values: [] }]
   },
 
   watch: {
-    value(value) {
+    value(value: string) {
       this.code = value;
       this.setValues();
     },
 
-    areaList: 'setValues'
+    areaList: 'setValues',
+
+    columnsNum(value: number) {
+      this.set({
+        displayColumns: this.data.columns.slice(0, +value)
+      });
+    }
   },
 
   methods: {
-    onCancel() {
-      this.$emit('cancel', {
-        values: this.getValues(),
-        indexs: this.getIndexs(),
-        detail: this.getDetail()
-      });
+    getPicker() {
+      if (this.picker == null) {
+        this.picker = this.selectComponent('.van-area__picker');
+      }
+      return this.picker;
     },
 
-    onConfirm() {
-      this.$emit('confirm', {
-        values: this.getValues(),
-        indexs: this.getIndexs(),
-        detail: this.getDetail()
-      });
+    onCancel(event: Weapp.Event) {
+      this.emit('cancel', event.detail);
+    },
+
+    onConfirm(event: Weapp.Event) {
+      this.emit('confirm', event.detail);
+    },
+
+    emit(type: string, detail) {
+      detail.values = detail.value;
+      delete detail.value;
+      this.$emit(type, detail);
     },
 
     onChange(event: Weapp.Event) {
-      const { value } = event.detail;
-      const { pickerValue, displayColumns } = this.data;
-      const index = pickerValue.findIndex(
-        (item, index) => item !== value[index]
-      );
-      const values = displayColumns[index];
-
-      if (index < 0 || value[index] < 0 || !values[value[index]]) {
-        return;
-      }
-
-      this.code = values[value[index]].code;
-      this.setValues();
-      this.$emit('change', {
-        picker: this,
-        values: this.getValues(),
-        index
+      const { index, picker, value } = event.detail;
+      this.code = value[index].code;
+      this.setValues().then(() => {
+        this.$emit('change', {
+          picker,
+          values: picker.getValues(),
+          index
+        });
       });
+    },
+
+    getConfig(type: string) {
+      const { areaList } = this.data;
+      return (areaList && areaList[`${type}_list`]) || {};
     },
 
     getList(type: string, code?: string): AreaItem[] {
@@ -93,7 +97,7 @@ VantComponent({
         return result;
       }
 
-      const list = this.data.areaList && this.data.areaList[`${type}_list`] || {};
+      const list = this.getConfig(type);
       result = Object.keys(list).map(code => ({
         code,
         name: list[code]
@@ -131,43 +135,48 @@ VantComponent({
     },
 
     setValues() {
-      let code =
-        this.code ||
-        (this.data.areaList &&
-          Object.keys(this.data.areaList.county_list || {})[0]) ||
-        '';
+      const county = this.getConfig('county');
+      let code = this.code || Object.keys(county)[0] || '';
       const province = this.getList('province');
       const city = this.getList('city', code.slice(0, 2));
 
-      this.set({
-        'columns[0]': province,
-        'columns[1]': city
-      });
+      const picker = this.getPicker();
 
-      if (city.length && code.slice(2, 4) === '00') {
-        code = city[0].code;
+      if (!picker) {
+        return;
       }
 
-      this.set({
-        'columns[2]': this.getList('county', code.slice(0, 4)),
-        pickerValue: [
-          this.getIndex('province', code),
-          this.getIndex('city', code),
-          this.getIndex('county', code)
-        ]
-      });
+      const stack = [];
+
+      stack.push(picker.setColumnValues(0, province, false));
+      stack.push(picker.setColumnValues(1, city, false));
+
+      if (city.length && code.slice(2, 4) === '00') {
+        ;[{ code }] = city;
+      }
+
+      stack.push(
+        picker.setColumnValues(
+          2,
+          this.getList('county', code.slice(0, 4)),
+          false
+        )
+      );
+
+      return Promise.all(stack)
+        .then(() =>
+          picker.setIndexes([
+            this.getIndex('province', code),
+            this.getIndex('city', code),
+            this.getIndex('county', code)
+          ])
+        )
+        .catch(() => {});
     },
 
     getValues() {
-      const { displayColumns = [], pickerValue = [] } = this.data;
-      return displayColumns
-        .map((option, index) => option[pickerValue[index]])
-        .filter(value => !!value);
-    },
-
-    getIndexs() {
-      const { pickerValue, columnsNum } = this.data;
-      return pickerValue.slice(0, columnsNum);
+      const picker = this.getPicker();
+      return picker ? picker.getValues().filter(value => !!value) : [];
     },
 
     getDetail() {
@@ -184,7 +193,7 @@ VantComponent({
         return area;
       }
 
-      const names = values.map(item => item.name);
+      const names = values.map((item: AreaItem) => item.name);
       area.code = values[values.length - 1].code;
       if (area.code[0] === '9') {
         area.country = names[1] || '';
@@ -200,7 +209,7 @@ VantComponent({
 
     reset() {
       this.code = '';
-      this.setValues();
+      return this.setValues();
     }
   }
 });