+++ packages/cell/index.ts
@@ -6,7 +6,8 @@ VantComponent({
     'title-class',
     'label-class',
     'value-class',
-    'right-icon-class'
+    'right-icon-class',
+    'hover-class'
   ],
 
   mixins: [link],
@@ -27,24 +28,6 @@ VantComponent({
     border: {
       type: Boolean,
       value: true
-    }
-  },
-
-  computed: {
-    cellClass(): string {
-      const { data } = this;
-      return this.classNames('van-cell', {
-        'van-cell--center': data.center,
-        'van-cell--required': data.required,
-        'van-cell--borderless': !data.border,
-        'van-cell--clickable': data.isLink || data.clickable,
-        [`van-cell--${data.size}`]: data.size
-      });
-    },
-
-    titleStyle(): string {
-      const { titleWidth } = this.data;
-      return titleWidth ? `max-width: ${titleWidth};min-width: ${titleWidth}` : '';
     }
   },