+++ packages/switch/index.ts
@@ -6,7 +6,7 @@ VantComponent({
   classes: ['node-class'],
 
   props: {
-    checked: Boolean,
+    checked: null,
     loading: Boolean,
     disabled: Boolean,
     activeColor: String,
@@ -14,6 +14,14 @@ VantComponent({
     size: {
       type: String,
       value: '30px'
+    },
+    activeValue: {
+      type: null,
+      value: true
+    },
+    inactiveValue: {
+      type: null,
+      value: false
     }
   },
 
@@ -23,30 +31,18 @@ VantComponent({
     }
   },
 
-  computed: {
-    classes(): string {
-      return this.classNames('van-switch', {
-        'van-switch--on': this.data.checked,
-        'van-switch--disabled': this.data.disabled
-      });
-    },
-
-    style() {
-      const backgroundColor = this.data.checked ? this.data.activeColor : this.data.inactiveColor;
-      return `font-size: ${this.data.size}; ${ backgroundColor ? `background-color: ${backgroundColor}` : '' }`
-    }
-  },
-
   created() {
     this.set({ value: this.data.checked });
   },
 
   methods: {
     onClick() {
+      const { activeValue, inactiveValue } = this.data;
       if (!this.data.disabled && !this.data.loading) {
-        const checked = !this.data.checked;
-        this.$emit('input', checked);
-        this.$emit('change', checked);
+        const checked = this.data.checked === activeValue;
+        const value = checked ? inactiveValue : activeValue;
+        this.$emit('input', value);
+        this.$emit('change', value);
       }
     }
   }