+++ packages/tabs/index.ts
@@ -2,11 +2,13 @@ import { VantComponent } from '../common/component';
 import { touch } from '../mixins/touch';
 
 type TabItemData = {
-  active: boolean;
-  inited?: boolean;
-  animated?: boolean;
-  width?: Number;
+  width?: number
+  active: boolean
+  inited?: boolean
+  animated?: boolean
 };
+
+type Position = 'top' | 'bottom' | '';
 
 VantComponent({
   mixins: [touch],
@@ -29,7 +31,14 @@ VantComponent({
 
   props: {
     color: String,
+    sticky: Boolean,
+    animated: Boolean,
+    swipeable: Boolean,
     lineWidth: {
+      type: Number,
+      value: -1
+    },
+    lineHeight: {
       type: Number,
       value: -1
     },
@@ -57,14 +66,7 @@ VantComponent({
       type: Number,
       value: 4
     },
-    animated: Boolean,
-    sticky: Boolean,
     offsetTop: {
-      type: Number,
-      value: 0
-    },
-    swipeable: Boolean,
-    scrollTop: {
       type: Number,
       value: 0
     }
@@ -88,9 +90,9 @@ VantComponent({
     },
     color: 'setLine',
     lineWidth: 'setLine',
+    lineHeight: 'setLine',
     active: 'setActiveTab',
     animated: 'setTrack',
-    scrollTop: 'onScroll',
     offsetTop: 'setWrapStyle'
   },
 
@@ -99,13 +101,22 @@ VantComponent({
   },
 
   mounted() {
-    this.setLine();
+    this.setLine(true);
     this.setTrack();
     this.scrollIntoView();
+
+    this.getRect('.van-tabs__wrap').then((rect: wx.BoundingClientRectCallbackResult) => {
+      this.navHeight = rect.height;
+      this.observerContentScroll();
+    });
+  },
+
+  destroyed() {
+    wx.createIntersectionObserver(this).disconnect();
   },
 
   methods: {
-    updateTabs(tabs) {
+    updateTabs(tabs: TabItemData[]) {
       tabs = tabs || this.data.tabs;
       this.set({
         tabs,
@@ -139,21 +150,17 @@ VantComponent({
       }
     },
 
-    setLine() {
+    setLine(skipTransition?: boolean) {
       if (this.data.type !== 'line') {
         return;
       }
 
-      const {
-        color,
-        active,
-        duration,
-        lineWidth
-      } = this.data;
+      const { color, active, duration, lineWidth, lineHeight } = this.data;
 
-      this.getRect('.van-tab', true).then(rects => {
+      this.getRect('.van-tab', true).then((rects: wx.BoundingClientRectCallbackResult[]) => {
         const rect = rects[active];
-        const width = (lineWidth !== -1) ? lineWidth : rect.width / 2;
+        const width = lineWidth !== -1 ? lineWidth : rect.width / 2;
+        const height = lineHeight !== -1 ? `height: ${lineHeight}px;` : '';
 
         let left = rects
           .slice(0, active)
@@ -161,53 +168,51 @@ VantComponent({
 
         left += (rect.width - width) / 2;
 
+        const transition = skipTransition
+          ? ''
+          : `transition-duration: ${duration}s; -webkit-transition-duration: ${duration}s;`;
+
         this.set({
           lineStyle: `
+            ${height}
             width: ${width}px;
             background-color: ${color};
             -webkit-transform: translateX(${left}px);
-            -webkit-transition-duration: ${duration}s;
             transform: translateX(${left}px);
-            transition-duration: ${duration}s;
+            ${transition}
           `
         });
       });
     },
 
     setTrack() {
-      const {
-        animated,
-        active,
-        duration
-      } = this.data;
+      const { animated, active, duration } = this.data;
 
       if (!animated) return '';
 
-      this.getRect('.van-tabs__content').then(rect => {
+      this.getRect('.van-tabs__content').then((rect: wx.BoundingClientRectCallbackResult) => {
         const { width } = rect;
 
         this.set({
           trackStyle: `
             width: ${width * this.child.length}px;
-            transform: translateX(${-1 * active * width}px);
-            transition-duration: ${duration}s;
+            left: ${-1 * active * width}px;
+            transition: left ${duration}s;
+            display: -webkit-box;
+            display: flex;
           `
         });
-        this.setTabsProps({
-          width,
-          animated
-        })
-      })
-    },
 
-    setTabsProps(props) {
-      this.child.forEach(item => {
-        item.set(props);
+        const props = { width, animated };
+
+        this.child.forEach((item: Weapp.Component) => {
+          item.set(props);
+        });
       });
     },
 
     setActiveTab() {
-      this.child.forEach((item, index) => {
+      this.child.forEach((item: Weapp.Component, index: number) => {
         const data: TabItemData = {
           active: index === this.data.active
         };
@@ -230,24 +235,27 @@ VantComponent({
 
     // scroll active tab into view
     scrollIntoView() {
-      if (!this.data.scrollable) {
+      const { active, scrollable } = this.data;
+
+      if (!scrollable) {
         return;
       }
 
-      this.getRect('.van-tab', true).then(tabRects => {
-        const tabRect = tabRects[this.data.active];
-        const offsetLeft = tabRects
-          .slice(0, this.data.active)
-          .reduce((prev, curr) => prev + curr.width, 0);
-        const tabWidth = tabRect.width;
+      Promise.all([
+        this.getRect('.van-tab', true),
+        this.getRect('.van-tabs__nav')
+      ]).then(
+        ([tabRects, navRect]: [wx.BoundingClientRectCallbackResult[], wx.BoundingClientRectCallbackResult]) => {
+          const tabRect = tabRects[active];
+          const offsetLeft = tabRects
+            .slice(0, active)
+            .reduce((prev, curr) => prev + curr.width, 0);
 
-        this.getRect('.van-tabs__nav').then(navRect => {
-          const navWidth = navRect.width;
           this.set({
-            scrollLeft: offsetLeft - (navWidth - tabWidth) / 2
+            scrollLeft: offsetLeft - (navRect.width - tabRect.width) / 2
           });
-        });
-      });
+        }
+      );
     },
 
     onTouchStart(event: Weapp.TouchEvent) {
@@ -281,8 +289,11 @@ VantComponent({
     },
 
     setWrapStyle() {
-      const { offsetTop, position } = this.data;
-      let wrapStyle;
+      const { offsetTop, position } = this.data as {
+        offsetTop: number
+        position: Position
+      };
+      let wrapStyle: string;
 
       switch (position) {
         case 'top':
@@ -304,44 +315,63 @@ VantComponent({
       // cut down `set`
       if (wrapStyle === this.data.wrapStyle) return;
 
-      this.set({
-        wrapStyle
-      });
+      this.set({ wrapStyle });
     },
 
-    // adjust tab position
-    onScroll(scrollTop) {
-      if (!this.data.sticky) return;
+    observerContentScroll() {
+      if (!this.data.sticky) {
+        return;
+      }
 
       const { offsetTop } = this.data;
+      const { windowHeight } = wx.getSystemInfoSync();
 
-      this.getRect('.van-tabs').then(rect => {
-        const { top, height } = rect;
+      wx.createIntersectionObserver(this)
+        .relativeToViewport({ top: -this.navHeight })
+        .observe('.van-tabs', res => {
+          const { top } = res.boundingClientRect;
 
-        this.getRect('.van-tabs__wrap').then(rect => {
-          const { height: wrapHeight } = rect;
-          let position = '';
-
-          if (offsetTop > top + height - wrapHeight) {
-            position = 'bottom';
-          } else if (offsetTop > top) {
-            position = 'top';
+          if (top > 0) {
+            return;
           }
 
+          const position: Position =
+            res.intersectionRatio > 0 ? 'top' : 'bottom';
+
           this.$emit('scroll', {
-            scrollTop: scrollTop + offsetTop,
+            scrollTop: top + offsetTop,
             isFixed: position === 'top'
           });
 
-          if (position !== this.data.position) {
-            this.set({
-              position
-            }, () => {
-              this.setWrapStyle();
-            });
-          }
+          this.setPosition(position);
         });
-      });
+
+      wx.createIntersectionObserver(this)
+        .relativeToViewport({ bottom: -(windowHeight - 1) })
+        .observe('.van-tabs', res => {
+          const { top, bottom } = res.boundingClientRect;
+
+          if (bottom < this.navHeight) {
+            return;
+          }
+
+          const position: Position = res.intersectionRatio > 0 ? 'top' : '';
+
+          this.$emit('scroll', {
+            scrollTop: top + offsetTop,
+            isFixed: position === 'top'
+          });
+
+          this.setPosition(position);
+        });
+    },
+
+    setPosition(position: Position) {
+      if (position !== this.data.position) {
+        this.set({ position }).then(() => {
+          this.setWrapStyle();
+        });
+      }
     }
   }
 });